Update Node.js versions in GitHub Actions due to v14 EOL
This commit updates the Node.js versions used in our GitHub Actions workflow in response to the EOL of Node.js v14.

Changes:
- Removed Node.js v14 from the testing matrix
- Now testing against Node.js v16 and v18

By removing the EOL version and focusing on v16 and v18, we can ensure that our application remains compatible with the latest and most stable releases of Node.js.